wip separating the Pure NN logic from the whole class
