+++ deeppavlov/models/go_bot/network.py
@@ -13,29 +13,70 @@
 # limitations under the License.
 
 # region imports
-import collections
 import json
 import re
 from logging import getLogger
-from typing import Dict, Any, List, Optional, Union, Tuple
+from typing import Dict, Any, List, Optional, Union
 
 import numpy as np
 import tensorflow as tf
-from tensorflow.contrib.layers import xavier_initializer as xav
 
 import deeppavlov.models.go_bot.templates as templ
 from deeppavlov import Chainer
 from deeppavlov.core.commands.utils import expand_path
 from deeppavlov.core.common.errors import ConfigError
 from deeppavlov.core.common.registry import register
-from deeppavlov.core.layers import tf_attention_mechanisms as am
-from deeppavlov.core.layers import tf_layers
 from deeppavlov.core.models.component import Component
 from deeppavlov.core.models.tf_model import LRScheduledTFModel
+from deeppavlov.models.go_bot.nn_stuff_handler import NNStuffHandler, configure_network_opts
 from deeppavlov.models.go_bot.tracker import FeaturizedTracker, DialogueStateTracker, MultipleUserStateTracker
 
 # endregion imports
 log = getLogger(__name__)
+
+
+def calc_obs_size(default_tracker_num_features,
+                  n_actions,
+                  bow_embedder, word_vocab, embedder,
+                  intent_classifier, intents):
+    obs_size = 6 + default_tracker_num_features + n_actions
+    if callable(bow_embedder):
+        obs_size += len(word_vocab)
+    if callable(embedder):
+        obs_size += embedder.dim
+    if callable(intent_classifier):
+        obs_size += len(intents)
+    log.info(f"Calculated input size for `GoalOrientedBotNetwork` is {obs_size}")
+    return obs_size
+
+
+def configure_attn(curr_attn_token_size,
+                   curr_attn_action_as_key,
+                   curr_attn_intent_as_key,
+                   curr_attn_key_size,
+                   embedder,
+                   n_actions,
+                   intent_classifier,
+                   intents):
+    token_size = curr_attn_token_size or embedder.dim
+    action_as_key = curr_attn_action_as_key or False
+    intent_as_key = curr_attn_intent_as_key or False
+
+    possible_key_size = 0
+    if action_as_key:
+        possible_key_size += n_actions
+    if intent_as_key and callable(intent_classifier):
+        possible_key_size += len(intents)
+    possible_key_size = possible_key_size or 1
+    key_size = curr_attn_key_size or possible_key_size
+
+    new_attn = {}
+    new_attn['token_size'] = token_size
+    new_attn['action_as_key'] = action_as_key
+    new_attn['intent_as_key'] = intent_as_key
+    new_attn['key_size'] = key_size
+
+    return new_attn
 
 
 @register("go_bot")
@@ -142,7 +183,7 @@ class GoalOrientedBot(LRScheduledTFModel):
                  debug: bool = False,
                  **kwargs) -> None:
 
-        self.nn_stuff_handler = self.NNStuffHandler()
+        self.nn_stuff_handler =     NNStuffHandler()
 
         # kwargs here are mostly training hyperparameters
 
@@ -211,48 +252,13 @@ class GoalOrientedBot(LRScheduledTFModel):
                       dense_size: int,
                       attn: dict) -> None:
         # initialize network
-        dense_size = dense_size or hidden_size
-        if obs_size is None:
-            obs_size = 6 + self.default_tracker.num_features + self.n_actions
-            if callable(self.bow_embedder):
-                obs_size += len(self.word_vocab)
-            if callable(self.embedder):
-                obs_size += self.embedder.dim
-            if callable(self.intent_classifier):
-                obs_size += len(self.intents)
-            log.info(f"Calculated input size for `GoalOrientedBotNetwork` is {obs_size}")
-        if action_size is None:
-            action_size = self.n_actions
-
-        if attn:
-            attn['token_size'] = attn.get('token_size') or self.embedder.dim
-            attn['action_as_key'] = attn.get('action_as_key', False)
-            attn['intent_as_key'] = attn.get('intent_as_key', False)
-
-            key_size = 0
-            if attn['action_as_key']:
-                key_size += self.n_actions
-            if attn['intent_as_key'] and callable(self.intent_classifier):
-                key_size += len(self.intents)
-            key_size = key_size or 1
-            attn['key_size'] = attn.get('key_size') or key_size
-
-        # specify model options
-        self.opt = {
-            'hidden_size': hidden_size,
-            'action_size': action_size,
-            'obs_size': obs_size,
-            'dense_size': dense_size,
-            'dropout_rate': dropout_rate,
-            'l2_reg_coef': l2_reg_coef,
-            'attention_mechanism': attn
-        }
+        self.nn_stuff_handler.configure_network_opts(action_size, attn, dense_size, dropout_rate, hidden_size, l2_reg_coef,
+                                     obs_size, self.embedder, self.n_actions, self.intent_classifier, self.intents,
+                                     self.default_tracker.num_features, self.bow_embedder, self.word_vocab)
 
         # initialize parameters
-        # self._init_network_params()
         self.nn_stuff_handler._init_network_params(self)
         # build computational graph
-        # self._build_graph()
         self.nn_stuff_handler._build_graph(self)
         # initialize session
         self.sess = tf.Session()
@@ -426,7 +432,8 @@ class GoalOrientedBot(LRScheduledTFModel):
         return b_features, b_emb_context, b_keys, b_u_masks, b_a_masks, b_actions
 
     def train_on_batch(self, x: List[dict], y: List[dict]) -> dict:
-        return self.network_train_on_batch(*self.prepare_data(x, y))
+        b_features, b_emb_context, b_keys, b_u_masks, b_a_masks, b_actions = self.prepare_data(x, y)
+        return self.network_train_on_batch(b_features, b_emb_context, b_keys, b_u_masks, b_a_masks, b_actions)
 
     # todo как инфер понимает из конфига что ему нужно. лёша что-то говорил про дерево
     def _infer(self, tokens: List[str], tracker: DialogueStateTracker) -> List:
@@ -509,8 +516,6 @@ class GoalOrientedBot(LRScheduledTFModel):
         if self.debug:
             log.debug("Bot reset.")
 
-    # region pure nn-specific stuff
-
     def network_call(self,
                      features: np.ndarray,
                      emb_context: np.ndarray,
@@ -563,317 +568,6 @@ class GoalOrientedBot(LRScheduledTFModel):
                 'learning_rate': self.get_learning_rate(),
                 'momentum': self.get_momentum()}
 
-    class NNStuffHandler():
-        def __init__(self):
-            pass
-
-        def _init_network_params(self, gobot_obj) -> None:
-            gobot_obj.dropout_rate = gobot_obj.opt['dropout_rate']
-            gobot_obj.hidden_size = gobot_obj.opt['hidden_size']
-            gobot_obj.action_size = gobot_obj.opt['action_size']
-            gobot_obj.obs_size = gobot_obj.opt['obs_size']
-            gobot_obj.dense_size = gobot_obj.opt['dense_size']
-            gobot_obj.l2_reg = gobot_obj.opt['l2_reg_coef']
-
-            attn = gobot_obj.opt.get('attention_mechanism')
-            if attn:
-                # gobot_obj.opt['attention_mechanism'] = attn
-
-                gobot_obj.attn = collections.namedtuple('attention_mechanism', attn.keys())(**attn)
-                gobot_obj.obs_size -= attn['token_size']
-            else:
-                gobot_obj.attn = None
-
-        def _build_graph(self, gobot_obj) -> None:
-
-            self._add_placeholders(gobot_obj)  # todo какая-то тензорфлововая тема
-
-            # build body
-            _logits, gobot_obj._state = self._build_body(gobot_obj)
-
-            # probabilities normalization : elemwise multiply with action mask
-            _logits_exp = tf.multiply(tf.exp(_logits), gobot_obj._action_mask)
-            _logits_exp_sum = tf.expand_dims(tf.reduce_sum(_logits_exp, -1), -1)
-            gobot_obj._probs = tf.squeeze(_logits_exp / _logits_exp_sum, name='probs')
-
-            # loss, train and predict operations
-            gobot_obj._prediction = tf.argmax(gobot_obj._probs, axis=-1, name='prediction')
-
-            # _weights = tf.expand_dims(self._utterance_mask, -1)
-            # TODO: try multiplying logits to action_mask
-            onehots = tf.one_hot(gobot_obj._action, gobot_obj.action_size)
-            _loss_tensor = tf.nn.softmax_cross_entropy_with_logits_v2(
-                logits=_logits, labels=onehots
-            )
-            # multiply with batch utterance mask
-            _loss_tensor = tf.multiply(_loss_tensor, gobot_obj._utterance_mask)
-            gobot_obj._loss = tf.reduce_mean(_loss_tensor, name='loss')
-            gobot_obj._loss += gobot_obj.l2_reg * tf.losses.get_regularization_loss()
-            gobot_obj._train_op = gobot_obj.get_train_op(gobot_obj._loss)
-
-        def _add_placeholders(self, gobot_obj) -> None:
-            # todo узнай что такое плейсхолдеры в тф
-            gobot_obj._dropout_keep_prob = tf.placeholder_with_default(1.0,
-                                                                       shape=[],
-                                                                       name='dropout_prob')
-            gobot_obj._features = tf.placeholder(tf.float32,
-                                                 [None, None, gobot_obj.obs_size],
-                                                 name='features')
-            gobot_obj._action = tf.placeholder(tf.int32,
-                                               [None, None],
-                                               name='ground_truth_action')
-            gobot_obj._action_mask = tf.placeholder(tf.float32,
-                                                    [None, None, gobot_obj.action_size],
-                                                    name='action_mask')
-            gobot_obj._utterance_mask = tf.placeholder(tf.float32,
-                                                       shape=[None, None],
-                                                       name='utterance_mask')
-            gobot_obj._batch_size = tf.shape(gobot_obj._features)[0]
-            zero_state = tf.zeros([gobot_obj._batch_size, gobot_obj.hidden_size], dtype=tf.float32)
-            _initial_state_c = \
-                tf.placeholder_with_default(zero_state, shape=[None, gobot_obj.hidden_size])
-            _initial_state_h = \
-                tf.placeholder_with_default(zero_state, shape=[None, gobot_obj.hidden_size])
-            gobot_obj._initial_state = tf.nn.rnn_cell.LSTMStateTuple(_initial_state_c,
-                                                                     _initial_state_h)
-            if gobot_obj.attn:
-                _emb_context_shape = \
-                    [None, None, gobot_obj.attn.max_num_tokens, gobot_obj.attn.token_size]
-                gobot_obj._emb_context = tf.placeholder(tf.float32,
-                                                        _emb_context_shape,
-                                                        name='emb_context')
-                gobot_obj._key = tf.placeholder(tf.float32,
-                                                [None, None, gobot_obj.attn.key_size],
-                                                name='key')
-
-        def _build_body(self, gobot_obj) -> Tuple[tf.Tensor, tf.Tensor]:
-            # input projection
-            _units = tf.layers.dense(gobot_obj._features, gobot_obj.dense_size,
-                                     kernel_regularizer=tf.nn.l2_loss,
-                                     kernel_initializer=xav())
-            if gobot_obj.attn:
-                attn_scope = f"attention_mechanism/{gobot_obj.attn.type}"
-                with tf.variable_scope(attn_scope):
-                    if gobot_obj.attn.type == 'general':
-                        _attn_output = am.general_attention(
-                            gobot_obj._key,
-                            gobot_obj._emb_context,
-                            hidden_size=gobot_obj.attn.hidden_size,
-                            projected_align=gobot_obj.attn.projected_align)
-                    elif gobot_obj.attn.type == 'bahdanau':
-                        _attn_output = am.bahdanau_attention(
-                            gobot_obj._key,
-                            gobot_obj._emb_context,
-                            hidden_size=gobot_obj.attn.hidden_size,
-                            projected_align=gobot_obj.attn.projected_align)
-                    elif gobot_obj.attn.type == 'cs_general':
-                        _attn_output = am.cs_general_attention(
-                            gobot_obj._key,
-                            gobot_obj._emb_context,
-                            hidden_size=gobot_obj.attn.hidden_size,
-                            depth=gobot_obj.attn.depth,
-                            projected_align=gobot_obj.attn.projected_align)
-                    elif gobot_obj.attn.type == 'cs_bahdanau':
-                        _attn_output = am.cs_bahdanau_attention(
-                            gobot_obj._key,
-                            gobot_obj._emb_context,
-                            hidden_size=gobot_obj.attn.hidden_size,
-                            depth=gobot_obj.attn.depth,
-                            projected_align=gobot_obj.attn.projected_align)
-                    elif gobot_obj.attn.type == 'light_general':
-                        _attn_output = am.light_general_attention(
-                            gobot_obj._key,
-                            gobot_obj._emb_context,
-                            hidden_size=gobot_obj.attn.hidden_size,
-                            projected_align=gobot_obj.attn.projected_align)
-                    elif gobot_obj.attn.type == 'light_bahdanau':
-                        _attn_output = am.light_bahdanau_attention(
-                            gobot_obj._key,
-                            gobot_obj._emb_context,
-                            hidden_size=gobot_obj.attn.hidden_size,
-                            projected_align=gobot_obj.attn.projected_align)
-                    else:
-                        raise ValueError("wrong value for attention mechanism type")
-                _units = tf.concat([_units, _attn_output], -1)
-
-            _units = tf_layers.variational_dropout(_units,
-                                                   keep_prob=gobot_obj._dropout_keep_prob)
-
-            # recurrent network unit
-            _lstm_cell = tf.nn.rnn_cell.LSTMCell(gobot_obj.hidden_size)
-            _utter_lengths = tf.cast(tf.reduce_sum(gobot_obj._utterance_mask, axis=-1),
-                                     tf.int32)
-            # _output: [batch_size, max_time, hidden_size]
-            # _state: tuple of two [batch_size, hidden_size]
-            _output, _state = tf.nn.dynamic_rnn(_lstm_cell,
-                                                _units,
-                                                time_major=False,
-                                                initial_state=gobot_obj._initial_state,
-                                                sequence_length=_utter_lengths)
-            _output = tf.reshape(_output, (gobot_obj._batch_size, -1, gobot_obj.hidden_size))
-            _output = tf_layers.variational_dropout(_output,
-                                                    keep_prob=gobot_obj._dropout_keep_prob)
-            # output projection
-            _logits = tf.layers.dense(_output, gobot_obj.action_size,
-                                      kernel_regularizer=tf.nn.l2_loss,
-                                      kernel_initializer=xav(), name='logits')
-            return _logits, _state
-
-    # def _init_network_params(self) -> None:
-    #     self.dropout_rate = self.opt['dropout_rate']
-    #     self.hidden_size = self.opt['hidden_size']
-    #     self.action_size = self.opt['action_size']
-    #     self.obs_size = self.opt['obs_size']
-    #     self.dense_size = self.opt['dense_size']
-    #     self.l2_reg = self.opt['l2_reg_coef']
-    #
-    #     attn = self.opt.get('attention_mechanism')
-    #     if attn:
-    #         self.opt['attention_mechanism'] = attn
-    #
-    #         self.attn = \
-    #             collections.namedtuple('attention_mechanism', attn.keys())(**attn)
-    #         self.obs_size -= attn['token_size']
-    #     else:
-    #         self.attn = None
-
-    # def _build_graph(self) -> None:
-    #
-    #     self._add_placeholders()  # todo какая-то тензорфлововая тема
-    #
-    #     # build body
-    #     _logits, self._state = self._build_body()
-    #
-    #     # probabilities normalization : elemwise multiply with action mask
-    #     _logits_exp = tf.multiply(tf.exp(_logits), self._action_mask)
-    #     _logits_exp_sum = tf.expand_dims(tf.reduce_sum(_logits_exp, -1), -1)
-    #     self._probs = tf.squeeze(_logits_exp / _logits_exp_sum, name='probs')
-    #
-    #     # loss, train and predict operations
-    #     self._prediction = tf.argmax(self._probs, axis=-1, name='prediction')
-    #
-    #     # _weights = tf.expand_dims(self._utterance_mask, -1)
-    #     # TODO: try multiplying logits to action_mask
-    #     onehots = tf.one_hot(self._action, self.action_size)
-    #     _loss_tensor = tf.nn.softmax_cross_entropy_with_logits_v2(
-    #         logits=_logits, labels=onehots
-    #     )
-    #     # multiply with batch utterance mask
-    #     _loss_tensor = tf.multiply(_loss_tensor, self._utterance_mask)
-    #     self._loss = tf.reduce_mean(_loss_tensor, name='loss')
-    #     self._loss += self.l2_reg * tf.losses.get_regularization_loss()
-    #     self._train_op = self.get_train_op(self._loss)
-
-    # def _add_placeholders(self) -> None:
-    #     # todo узнай что такое плейсхолдеры в тф
-    #     self._dropout_keep_prob = tf.placeholder_with_default(1.0,
-    #                                                           shape=[],
-    #                                                           name='dropout_prob')
-    #     self._features = tf.placeholder(tf.float32,
-    #                                     [None, None, self.obs_size],
-    #                                     name='features')
-    #     self._action = tf.placeholder(tf.int32,
-    #                                   [None, None],
-    #                                   name='ground_truth_action')
-    #     self._action_mask = tf.placeholder(tf.float32,
-    #                                        [None, None, self.action_size],
-    #                                        name='action_mask')
-    #     self._utterance_mask = tf.placeholder(tf.float32,
-    #                                           shape=[None, None],
-    #                                           name='utterance_mask')
-    #     self._batch_size = tf.shape(self._features)[0]
-    #     zero_state = tf.zeros([self._batch_size, self.hidden_size], dtype=tf.float32)
-    #     _initial_state_c = \
-    #         tf.placeholder_with_default(zero_state, shape=[None, self.hidden_size])
-    #     _initial_state_h = \
-    #         tf.placeholder_with_default(zero_state, shape=[None, self.hidden_size])
-    #     self._initial_state = tf.nn.rnn_cell.LSTMStateTuple(_initial_state_c,
-    #                                                         _initial_state_h)
-    #     if self.attn:
-    #         _emb_context_shape = \
-    #             [None, None, self.attn.max_num_tokens, self.attn.token_size]
-    #         self._emb_context = tf.placeholder(tf.float32,
-    #                                            _emb_context_shape,
-    #                                            name='emb_context')
-    #         self._key = tf.placeholder(tf.float32,
-    #                                    [None, None, self.attn.key_size],
-    #                                    name='key')
-
-    # def _build_body(self) -> Tuple[tf.Tensor, tf.Tensor]:
-    #     # input projection
-    #     _units = tf.layers.dense(self._features, self.dense_size,
-    #                              kernel_regularizer=tf.nn.l2_loss,
-    #                              kernel_initializer=xav())
-    #     if self.attn:
-    #         attn_scope = f"attention_mechanism/{self.attn.type}"
-    #         with tf.variable_scope(attn_scope):
-    #             if self.attn.type == 'general':
-    #                 _attn_output = am.general_attention(
-    #                     self._key,
-    #                     self._emb_context,
-    #                     hidden_size=self.attn.hidden_size,
-    #                     projected_align=self.attn.projected_align)
-    #             elif self.attn.type == 'bahdanau':
-    #                 _attn_output = am.bahdanau_attention(
-    #                     self._key,
-    #                     self._emb_context,
-    #                     hidden_size=self.attn.hidden_size,
-    #                     projected_align=self.attn.projected_align)
-    #             elif self.attn.type == 'cs_general':
-    #                 _attn_output = am.cs_general_attention(
-    #                     self._key,
-    #                     self._emb_context,
-    #                     hidden_size=self.attn.hidden_size,
-    #                     depth=self.attn.depth,
-    #                     projected_align=self.attn.projected_align)
-    #             elif self.attn.type == 'cs_bahdanau':
-    #                 _attn_output = am.cs_bahdanau_attention(
-    #                     self._key,
-    #                     self._emb_context,
-    #                     hidden_size=self.attn.hidden_size,
-    #                     depth=self.attn.depth,
-    #                     projected_align=self.attn.projected_align)
-    #             elif self.attn.type == 'light_general':
-    #                 _attn_output = am.light_general_attention(
-    #                     self._key,
-    #                     self._emb_context,
-    #                     hidden_size=self.attn.hidden_size,
-    #                     projected_align=self.attn.projected_align)
-    #             elif self.attn.type == 'light_bahdanau':
-    #                 _attn_output = am.light_bahdanau_attention(
-    #                     self._key,
-    #                     self._emb_context,
-    #                     hidden_size=self.attn.hidden_size,
-    #                     projected_align=self.attn.projected_align)
-    #             else:
-    #                 raise ValueError("wrong value for attention mechanism type")
-    #         _units = tf.concat([_units, _attn_output], -1)
-    #
-    #     _units = tf_layers.variational_dropout(_units,
-    #                                            keep_prob=self._dropout_keep_prob)
-    #
-    #     # recurrent network unit
-    #     _lstm_cell = tf.nn.rnn_cell.LSTMCell(self.hidden_size)
-    #     _utter_lengths = tf.cast(tf.reduce_sum(self._utterance_mask, axis=-1),
-    #                              tf.int32)
-    #     # _output: [batch_size, max_time, hidden_size]
-    #     # _state: tuple of two [batch_size, hidden_size]
-    #     _output, _state = tf.nn.dynamic_rnn(_lstm_cell,
-    #                                         _units,
-    #                                         time_major=False,
-    #                                         initial_state=self._initial_state,
-    #                                         sequence_length=_utter_lengths)
-    #     _output = tf.reshape(_output, (self._batch_size, -1, self.hidden_size))
-    #     _output = tf_layers.variational_dropout(_output,
-    #                                             keep_prob=self._dropout_keep_prob)
-    #     # output projection
-    #     _logits = tf.layers.dense(_output, self.action_size,
-    #                               kernel_regularizer=tf.nn.l2_loss,
-    #                               kernel_initializer=xav(), name='logits')
-    #     return _logits, _state
-
-    # endregion pure nn-specific stuff
-
     # region helping stuff
     def load(self, *args, **kwargs) -> None:
         self.load_params()
@@ -887,7 +581,7 @@ class GoalOrientedBot(LRScheduledTFModel):
         path = str(self.save_path.with_suffix('.json').resolve())
         log.info(f"[saving parameters to {path}]")
         with open(path, 'w', encoding='utf8') as fp:
-            json.dump(self.opt, fp)
+            json.dump(self.nn_stuff_handler.opt, fp)
 
     def load_params(self) -> None:
         path = str(self.load_path.with_suffix('.json').resolve())
@@ -895,10 +589,10 @@ class GoalOrientedBot(LRScheduledTFModel):
         with open(path, 'r', encoding='utf8') as fp:
             params = json.load(fp)
         for p in self.GRAPH_PARAMS:
-            if self.opt.get(p) != params.get(p):
+            if self.nn_stuff_handler.opt.get(p) != params.get(p):
                 raise ConfigError(f"`{p}` parameter must be equal to saved"
                                   f" model parameter value `{params.get(p)}`,"
-                                  f" but is equal to `{self.opt.get(p)}`")
+                                  f" but is equal to `{self.nn_stuff_handler.opt.get(p)}`")
 
     def process_event(self, event_name, data) -> None:
         super().process_event(event_name, data)
